update: added priority to a post
