+++ backend/middlewares/authMiddleware.ts
@@ -1,6 +1,18 @@
 import { NextFunction, Request, Response } from "express";
 import jwt, { JwtPayload } from "jsonwebtoken";
-import { User } from "../types";
+import { ObjectId } from "mongoose";
+
+export interface User {
+  _id: ObjectId;
+  username: string;
+  password: string;
+  firstname: string;
+  lastname: string;
+  address: string;
+  posts: [];
+  iat: number;
+  exp: number;
+}
 
 export interface AuthRequest extends Request {
   // user: JwtPayload;
@@ -30,11 +42,6 @@ const authMiddleware = (
     }
 
     req.user = decoded.user;
-
-    console.log(
-      "decodeduserdecodeduserdecodeduserdecodeduserdecodeduserdecodeduser",
-      decoded.user
-    );
 
     return next();
   } catch (error) {
